# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Piano Chord Learner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Song {
  id: number;
  title: string;
  /** @nullable */
  artist?: string | null;
  /** Raw chord progression text pasted by the user (e.g. "Am C G F") */
  chordText: string;
  createdAt: string;
}

export interface SongInput {
  /** @minLength 1 */
  title: string;
  artist?: string;
  /** @minLength 1 */
  chordText: string;
}

export interface SongUpdate {
  /** @minLength 1 */
  title?: string;
  artist?: string;
  /** @minLength 1 */
  chordText?: string;
}