# lib/api-zod/src/generated/types/song.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Piano Chord Learner API
 * OpenAPI spec version: 0.1.0
 */

export interface Song {
  id: number;
  title: string;
  /** @nullable */
  artist?: string | null;
  /** Raw chord progression text pasted by the user (e.g. "Am C G F") */
  chordText: string;
  createdAt: Date;
}