# lib/api-zod/src/generated/types/songInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Piano Chord Learner API
 * OpenAPI spec version: 0.1.0
 */

export interface SongInput {
  /** @minLength 1 */
  title: string;
  artist?: string;
  /** @minLength 1 */
  chordText: string;
}